uses express, simplifies server logic
