+++ package.json
@@ -48,10 +48,10 @@
   },
   "dependencies": {
     "commander": "^2.9.0",
+    "express": "^4.15.2",
     "git-url-parse": "^6.2.2",
     "mongoose": "^4.9.4",
-    "openbadge": "^1.0.4",
-    "path-to-regexp": "^1.7.0"
+    "openbadge": "^1.0.4"
   },
   "repository": {
     "type": "git",
